Refactored client and repository modules
